Added flash-only logging mode
Two logging modes now possible - Flash only and traditional (SD+Flash). Also this code does not use an address pointer for the Flash memory. On startup the last flash address is located and the address is just held in volatile memory. An erase of the flash memory is now a true erase.